chore: refactor chat api helper functions
