+++ app/ai_prompt/components/input-bar.tsx
@@ -1,6 +1,7 @@
-import { FC, FormEvent, useState } from 'react';
+import { FC, FormEvent } from 'react';
 import { Send } from 'lucide-react';
 import { Button } from '@/components/ui/button';
+import { Textarea } from '@/components/ui/textarea';
 
 interface Props {
   input: string;
@@ -8,25 +9,27 @@ interface Props {
   handleSubmit: (e: FormEvent<HTMLFormElement>) => void; // Accepts a FormEvent
 }
 
-export const InputBar: FC<Props> = ({ input, handleInputChange, handleSubmit }) => {
+export const InputBar: FC<Props> = ({
+  input,
+  handleInputChange,
+  handleSubmit,
+}) => {
   return (
     <div className="p-4 pt-0 w-full">
-      <div className="bg-foreground/5 border rounded-lg p-2 flex items-center">
-        <form onSubmit={handleSubmit}>
-          <div className="flex flex-col justify-between w-full">
-            <textarea
-              className="bg-gray-600/0 w-full h-18 text-white focus:outline-none overflow-hidden text-md resize-none"
-              value={input}
-              onChange={handleInputChange}
-              placeholder="Prompt your AI here..."
-            />
+      <form onSubmit={handleSubmit} className="w-full">
+        <div className="flex gap-2 w-full">
+          <Textarea
+            className="w-full flex-1 bg-foreground/5 border rounded-lg p-2 h-18 focus:outline-none overflow-hidden text-md resize-none"
+            value={input}
+            onChange={handleInputChange}
+            placeholder="Prompt your AI here..."
+          />
 
-            <Button type="submit" variant="ghost" size="icon">
-              <Send className="w-6 h-6" />
-            </Button>
-          </div>
-        </form>
-      </div>
+          <Button type="submit" variant="ghost" size="icon">
+            <Send className="w-6 h-6" />
+          </Button>
+        </div>
+      </form>
     </div>
   );
 };
